move manipulator to doghouse
