+++ src/main/java/frc/robot/RobotContainer.java
@@ -13,7 +13,10 @@
 
 package frc.robot;
 
-import static frc.robot.subsystems.vision.VisionConstants.*;
+import static frc.robot.subsystems.vision.VisionConstants.camera0Name;
+import static frc.robot.subsystems.vision.VisionConstants.camera1Name;
+import static frc.robot.subsystems.vision.VisionConstants.robotToCamera0;
+import static frc.robot.subsystems.vision.VisionConstants.robotToCamera1;
 
 import com.pathplanner.lib.auto.AutoBuilder;
 import com.pathplanner.lib.auto.NamedCommands;
@@ -26,6 +29,7 @@ import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.Commands;
 import edu.wpi.first.wpilibj2.command.ParallelCommandGroup;
+import edu.wpi.first.wpilibj2.command.ParallelDeadlineGroup;
 import edu.wpi.first.wpilibj2.command.WaitCommand;
 import edu.wpi.first.wpilibj2.command.button.CommandGenericHID;
 import edu.wpi.first.wpilibj2.command.button.CommandXboxController;
@@ -71,13 +75,6 @@ public class RobotContainer {
   public Doghouse m_doghouse = new Doghouse();
 
   public Climber climber = new Climber();
-
-  public Command coralMode() {
-
-    return new ParallelCommandGroup(
-        m_doghouse.moveFastFactory(), m_armevator.manipulatorInFactory());
-    // .until(m_armevator.isIn());
-  }
 
   public RobotContainer() {
     switch (Constants.currentMode) {
@@ -156,8 +153,8 @@ public class RobotContainer {
   }
 
   public void registerNamedCommands() {
-    NamedCommands.registerCommand("Shoot", m_armevator.manipulatorOutFactory());
-    NamedCommands.registerCommand("Get Coral", coralMode());
+    NamedCommands.registerCommand("Shoot", m_doghouse.manipulatorOutFactory());
+    NamedCommands.registerCommand("Get Coral", getCoralFactory());
     NamedCommands.registerCommand(
         "Arm Go To Zero", m_armevator.armGoToZeroFactory().withTimeout(0.5));
   }
@@ -177,16 +174,16 @@ public class RobotContainer {
     SmartDashboard.putData("Set Position Arm", m_armevator.setArmPostionFactory());
     SmartDashboard.putData("Arm Go To Zero", m_armevator.armGoToZeroFactory());
     SmartDashboard.putData("Stop Arm", m_armevator.stopArmFactory());
-    SmartDashboard.putData("Out Manipulator", m_armevator.manipulatorOutFactory());
-    SmartDashboard.putData("In Manipulator", m_armevator.manipulatorInFactory());
-    SmartDashboard.putData("Stop Manipulator", m_armevator.manipulatorStopFactory());
+    SmartDashboard.putData("Out Manipulator", m_doghouse.manipulatorOutFactory());
+    SmartDashboard.putData("In Manipulator", m_doghouse.manipulatorInFactory());
+    SmartDashboard.putData("Stop Manipulator", m_doghouse.manipulatorStopFactory());
     SmartDashboard.putData("Go To One Inch", m_armevator.goToOneInchFactory());
-    SmartDashboard.putData("Go To Tilt Angle", m_armevator.goToTiltAngleFactory());
-    SmartDashboard.putData("Algae Pickup", m_armevator.algaePickupFactory());
+    SmartDashboard.putData("Go To Tilt Angle", goToTiltAngleFactory());
+    SmartDashboard.putData("Algae Pickup", algaePickupFactory());
 
     SmartDashboard.putData("Stop Doghouse", m_doghouse.stopMovingFactory());
     SmartDashboard.putData("Slow Doghouse", m_doghouse.moveSlowFactory());
-    SmartDashboard.putData("Fast Doghouse", coralMode());
+    SmartDashboard.putData("Get Coral", getCoralFactory());
 
     SmartDashboard.putData("L4", m_armevator.L4Factory());
     SmartDashboard.putData("L3", m_armevator.L3Factory());
@@ -215,13 +212,13 @@ public class RobotContainer {
     buttons.button(1).onTrue(m_armevator.L2Factory());
     buttons.button(2).onTrue(m_armevator.L3Factory());
     buttons.button(3).onTrue(m_armevator.L4Factory());
-    buttons.button(10).onTrue(m_armevator.manipulatorOutFactory());
+    buttons.button(10).onTrue(m_doghouse.manipulatorOutFactory());
     buttons.button(4).onTrue(m_armevator.stowFactory());
-    buttons.button(5).onTrue(coralMode());
-    buttons.button(11).onTrue(m_armevator.algaePickupFactory());
+    buttons.button(5).onTrue(getCoralFactory());
+    buttons.button(11).onTrue(algaePickupFactory());
     buttons.button(7).onTrue(climber.prepClimber());
 
-    controller.rightBumper().onTrue(m_armevator.manipulatorOutFactory());
+    controller.rightBumper().onTrue(m_doghouse.manipulatorOutFactory());
   }
 
   /**
@@ -276,6 +273,20 @@ public class RobotContainer {
       e.printStackTrace();
       return autoPath;
     }
+  }
+
+  private Command algaePickupFactory() {
+    return new ParallelCommandGroup(
+        m_armevator.algaePickupFactory(), m_doghouse.algaePickupFactory());
+  }
+
+  private Command goToTiltAngleFactory() {
+    return new ParallelCommandGroup(
+        m_armevator.goToTiltAngleFactory(), m_doghouse.manipulatorStopFactory());
+  }
+
+  private Command getCoralFactory() {
+    return new ParallelDeadlineGroup(m_doghouse.coralIntake(), m_armevator.armGoToZeroFactory());
   }
 
   public void teleopInit() {}
